requirements.txt
# Core pipeline
pipecat-ai[openai,silero]
pipecat-ai[sarvam]
pipecat-ai[websocket]

# PIOPIY SDK - Install with: pip install "piopiy-ai[openai,deepgram,cartesia,silero,groq,sarvam]"
# This includes support for all STT/LLM/TTS providers (Deepgram, OpenAI, Cartesia, Groq, Sarvam)
piopiy-ai[openai,deepgram,cartesia,silero,groq,sarvam]

# Backend
fastapi
uvicorn
celery
redis
python-dotenv
requests
httpx
aiohttp
groq
PyJWT
python-multipart

# Utilities
loguru
websockets

# Follow-up automation
yagmail  # Email sending
pytz     # Timezone handling
